Move exception method
